Rework backend importing for timeline
- Fail with useful error if neither backend imports
- Add command-line arg to specify backend explicitly
